+++ libraries/ui-library-angular/src/lib/link/six-router-link.directive.ts
@@ -1,15 +1,5 @@
-import {
-  Attribute,
-  Directive,
-  ElementRef,
-  Input,
-  OnChanges,
-  OnInit,
-  Optional,
-  Renderer2,
-  SimpleChanges,
-} from '@angular/core';
-import { ActivatedRoute, QueryParamsHandling, Router, RouterLink } from '@angular/router';
+import { Directive, ElementRef, HostListener, Input, OnChanges, OnInit, Optional, Renderer2 } from '@angular/core';
+import { Router, RouterLink } from '@angular/router';
 import { LocationStrategy } from '@angular/common';
 
 /**
@@ -23,8 +13,15 @@ export class SixRouterLinkDirective implements OnInit, OnChanges {
     private locationStrategy: LocationStrategy,
     private elementRef: ElementRef,
     private router: Router,
+    private renderer: Renderer2,
     @Optional() private routerLinkDirective?: RouterLink
   ) {}
+
+  @HostListener('click', ['$event'])
+  onClick(event: UIEvent): void {
+    // Prevents the browser from performing a page reload when pressing a SIX-component with routerLink.
+    event.preventDefault();
+  }
 
   @Input() routerLink?: any;
   @Input() queryParams?: any;
@@ -42,9 +39,8 @@ export class SixRouterLinkDirective implements OnInit, OnChanges {
 
   private updateTargetUrlAndHref(): void {
     if (this.routerLinkDirective?.urlTree) {
-      this.elementRef.nativeElement.href = this.locationStrategy.prepareExternalUrl(
-        this.router.serializeUrl(this.routerLinkDirective.urlTree)
-      );
+      const url = this.locationStrategy.prepareExternalUrl(this.router.serializeUrl(this.routerLinkDirective.urlTree));
+      this.renderer.setAttribute(this.elementRef.nativeElement, 'href', url);
     }
   }
 }